refactor: move interfaces from src/server to src/shared
