+++ src/server/index.ts
@@ -1,5 +1,4 @@
 export * from './api';
 export * from './database';
-export * from './interfaces';
 export * from './services';
 export * from './utils';
