+++ src/shared/interfaces/chat-client.ts
