+++ src/shared/interfaces/config.ts
