+++ src/shared/interfaces/index.ts
@@ -1,3 +1,4 @@
 export type * from './chat-client';
 export type * from './config';
 export type * from './simpletex';
+export type * from './tour-step';
