+++ src/shared/interfaces/simpletex.ts
