+++ src/shared/interfaces/tour-step.ts
